+++ .gitignore
@@ -4,7 +4,7 @@
 __pycache__
 *.pyc
 .mypy_cache
-iohtap.egg-info
+brad.egg-info
 
 .DS_Store
 ._.DS_Store
