+++ src/brad/__init__.py
@@ -1,0 +1,7 @@
+__version__ = "0.0.0+dev"
+
+__author__ = "BRAD Authors"
+__email__ = "iohtap@csail.mit.edu"
+
+__description__ = "BRAD: A cloud-native multi-engine database management system."
+__url__ = "http://dsg.csail.mit.edu/"
