+++ src/brad/config/file.py
@@ -1,0 +1,101 @@
+import yaml
+
+from brad.config.dbtype import DBType
+from brad.config.extraction import ExtractionStrategy
+from brad.config.routing_policy import RoutingPolicy
+
+
+class ConfigFile:
+    def __init__(self, path: str):
+        with open(path, "r", encoding="UTF-8") as file:
+            self._raw = yaml.load(file, Loader=yaml.Loader)
+
+    @property
+    def server_interface(self) -> str:
+        return self._raw["server_interface"]
+
+    @property
+    def server_port(self) -> int:
+        return int(self._raw["server_port"])
+
+    @property
+    def server_daemon_port(self) -> int:
+        return int(self._raw["server_daemon_port"])
+
+    @property
+    def athena_s3_data_path(self) -> str:
+        return _ensure_slash_terminated(self._raw[DBType.Athena]["s3_data_path"])
+
+    @property
+    def extraction_strategy(self) -> ExtractionStrategy:
+        return ExtractionStrategy.from_str(self._raw["extraction_strategy"])
+
+    @property
+    def redshift_s3_iam_role(self) -> str:
+        """Needed when importing data from S3."""
+        return self._raw[DBType.Redshift]["s3_iam_role"]
+
+    @property
+    def s3_extract_bucket(self) -> str:
+        """Needed when exporting data from Aurora to S3."""
+        return self._raw["s3_extract_bucket"]
+
+    @property
+    def s3_extract_path(self) -> str:
+        """Needed when exporting data from Aurora to S3."""
+        return _ensure_slash_terminated(self._raw["s3_extract_path"])
+
+    @property
+    def s3_extract_region(self) -> str:
+        """Needed when exporting data from Aurora to S3."""
+        return self._raw["s3_extract_region"]
+
+    @property
+    def data_sync_period_seconds(self) -> float:
+        return float(self._raw["data_sync_period_seconds"])
+
+    @property
+    def routing_policy(self) -> RoutingPolicy:
+        return RoutingPolicy.from_str(self._raw["routing_policy"])
+
+    def get_odbc_connection_string(self, db: DBType) -> str:
+        if db not in self._raw:
+            raise AssertionError("Unhandled database type: " + str(db))
+
+        config = self._raw[db]
+        if db is DBType.Athena:
+            return "Driver={{{}}};AwsRegion={};S3OutputLocation={};AuthenticationType=IAM Credentials;UID={};PWD={};Schema={};".format(
+                config["odbc_driver"],
+                config["aws_region"],
+                config["s3_output_path"],
+                config["access_key"],
+                config["access_key_secret"],
+                config["database"],
+            )
+
+        elif db is DBType.Aurora:
+            return "Driver={{{}}};Server={};Port={};Uid={};Pwd={};Database={};".format(
+                config["odbc_driver"],
+                config["host"],
+                config["port"],
+                config["user"],
+                config["password"],
+                config["database"],
+            )
+
+        elif db is DBType.Redshift:
+            return "Driver={{{}}};Server={};Port={};Uid={};Pwd={};Database={};".format(
+                config["odbc_driver"],
+                config["host"],
+                config["port"],
+                config["user"],
+                config["password"],
+                config["database"],
+            )
+
+
+def _ensure_slash_terminated(candidate: str) -> str:
+    if not candidate.endswith("/"):
+        return candidate + "/"
+    else:
+        return candidate
