+++ src/brad/forecasting/query_parser.py
