--- src/iohtap/__init__.py
@@ -1,7 +1,0 @@
-__version__ = "0.0.0+dev"
-
-__author__ = "IOHTAP Authors"
-__email__ = "iohtap@csail.mit.edu"
-
-__description__ = "IOHTAP: A cloud-native federated database management system."
-__url__ = "http://dsg.csail.mit.edu/"
