--- src/iohtap/forecasting/README.md
@@ -1,1 +1,0 @@
-### This subdirectory hold code and other artifacts related to workload forecasting for iohtap
